changeslow update
I've gotten changeslow.c to correctly generate the minimum number of
coins required, but I cannot generate the array of which coins those
need to be. The function I currently have in the code works for the
first case in test.txt, but not the second or third cases. Test.txt
contains the three example problems laid out in the project document.